adds #side_worker to DSL
The `#side_worker` cannot change the flow of products, but can do work
to create side effects.  It's worth observing that any worker _could_
create side effects, but that should generally be discouraged.  This
worker has been added as a means to do side-effect work in a more
intention-revealing way.

Also, the side_worker's callable is not called if the worker is handed a
nil.  It's possible that a user might want the callable invoked for
after calls, but the errors which could result from an unexpected,
unhandled nil could be nasty to track down.  Perhaps we can consider
adding a `#side_worker!` variant later which executes even when the
value handed in is nil.